refactor([issue-2532]): split Universe Builder lifecycles
