+++ client/src/hooks/index.js
@@ -46,7 +46,11 @@ export { default as useRowDraft } from './useRowDraft.js';
 export { default as useTheme } from './useTheme.js';
 export { default as useUniverse } from './useUniverse.js';
 export { default as useUniverseAction } from './useUniverseAction.js';
+export { default as useUniverseDraft } from './useUniverseDraft.js';
+export * from './useUniverseDraft.js';
+export { default as useUniverseExpand } from './useUniverseExpand.js';
 export { default as useUniverseGallery } from './useUniverseGallery.js';
+export { default as useUniverseRender } from './useUniverseRender.js';
 export { default as useVideoDownload } from './useVideoDownload.js';
 export { default as useYoutubeTrackImport } from './useYoutubeTrackImport.js';
 export { default as useReferenceAudioImport } from './useReferenceAudioImport.js';
